requirements.txt
discord.py>=1.5.1
python-Levenshtein-wheels>=0.13.1
aiohttp>=3.6.2
asyncpg>=0.20.1
python-dotenv>=0.13.0
yoyo-migrations>=7.0.2
psycopg2-binary>=2.8.5
yarl==1.4.2
steam
iso3166